requirements.txt
yt-dlp
openai-whisper
youtube-transcript-api
static-ffmpeg
